Update README from JSON (source)
These changes will update the README from JSON file using GitHub actions. The JSON file becomes the source of data for the README. The README will be updated automatically when the JSON file is updated. The JSON file will be used to generate the README file using a python script. The python script will be run using GitHub actions.